Remove helpers/postgres_db file
The only function in this file is used to connect to a PostgreSQL when a
database isn't specified. I believe this functionality should be moved
to the db package going forwards as it is quite similar to its
`DBConnect()` function.
Variable names relating to the database connection have had to be
changed as part of this commit for the sake of clarity.