Move SQL functions to db/operations.go
These functions perform operations in a PostgreSQL database and as such
probably make more sense in the db package. I have also moved the code
used to create and drop the database to the db/operations.go file.
The public functions in this file will now return error values, making
them easier to test. The `db.SetSqlColumns()` function has been removed
from this package as I believe that it is only relevant to application
code and so is unecessary in the CLI.