Remove the bun package
I've had a great deal of difficulty integrating the migration package in
`bun` package into this project. This commit reverts some of the changes
made in the commit that introduced `bun` however, as `go-pg` is
deprecated, I will use Go's `database/sql` package (and the pq driver)
to connect to Postgresql going forwards.