Refactor db Operations tests
Prior to the changes in this commit it wasn't possible to run the test
suite without running into the following error:

    pq: database "pgcli_test" is being accessed by other users

As mocking the database connection proved quite difficult to do with
established mocking libraries[1] due to the way in which golang-migrate
perform database queries, it was necessary to deal with the issue
head-on. This commit ensures that the database is and dropped after each
test and any other user sessions in the database are terminated at this
point.

[1]: https://github.com/DATA-DOG/go-sqlmock/issues/239