Add migration functions and refactor tests
In order to adequately test the migration functions, I have had to
refactor the somewhat so that they now return error values (similar to
the operations functions in previous tests).
This commit also removes the `migrationHelper()` function as (following
the addition of the migration tests) I felt that this could now be
replaced with the existing `MigrateUp()` function in the operations
tests.