Add tests for DB connections helpers
The test coverage report showed that the `sslMode()` function had a
relatively low coverage so I have decided to add tests for this as well
as the `host()` and `port()` functions. Adding the tests revealed a bug
in the `sslMode()` function that would set the mode to `disable`
regardless of a user's configuration settings. This commit fixes this
issue as well.