Remove "relative" imports
This commit reverts changes made in a previous PR [1] which attempted to
use relative imports within the module. I'm quite new to go modules and
unfortunately that change was equivalent to renaming the module -
something that golang doesn't support.[2] This commit reverts the
changes and adds warnings to previous releases that were published in
error.

[1]: https://github.com/kohrVid/pg-cli/pull/3/commits/e05c9d229c668bcbffea996c1cceb35fed73e969
[2]: https://github.com/golang/go/issues/26904