Refactor the step command
Instructions in the output of the `step` command were incorrect and would
have led users to migrate to a specific version number instead of
"stepping through" migration versions. This commit fixes the usage
instructions but also refactors the command so that an `-n` flag
("number of steps") or `--` can be used instead of the repetitively
named `--step` flag.