Removing CircleCI configuration
I'm removing the CircleCI configuration from this project because of the
organisation's cowardly response to the genocide taking place in
Palestine. [1] They are far from the only tech company that has behaved
appallingly over this issue but their product is easy enough to remove
in this instance.

Unfortunately I do not have time to replace it with anything else as
v1.0.0 is blocking another project that I'm working on. For now, anyone
who uses this package will just have to trust that the tests are passing
on my machine!

[1]: https://news.ycombinator.com/item?id=38744050